requirements.txt
paho-mqtt
numpy
matplotlib
streamlit
pandas
